requirements.txt
Flask==1.1.2
pyTelegramBotAPI==3.7.6
gspread==3.7.0
oauth2client==4.1.3
Pillow==8.1.2
